Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -20,9 +20,9 @@ public:
         {
             auto topo = q.front();
             
-            int l = q.front().second;
+            long long l = q.front().second;
           
-            int r = q.back().second;
+           long long  r = q.back().second;
             maxw = max(maxw, r-l+1);
             int n = q.size();
             while(n--){
